requirements.txt
antlr4-python3-runtime==4.13.1
numpy==1.26.4
sympy==1.12
torch==2.8.0
requests==2.31.0
Flask==3.0.0
SQLAlchemy==2.0.25
